backend/pyproject.toml
[project]
name = "drivechill"
version = "1.0.0"
description = "PC fan controller based on hardware temperatures"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiosqlite>=0.19.0",
    "psutil>=5.9.0",
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
    "httpx>=0.26.0",
    "bcrypt>=4.0.0",
]

[project.optional-dependencies]
linux = ["pySMART>=1.3.0"]
dev = ["pytest>=7.4.0", "pytest-asyncio>=0.23.0", "ruff>=0.1.0"]

[tool.ruff]
line-length = 100
target-version = "py311"
